fix(protocol): send setTxPower as 1 byte instead of 4
- Changed PacketBuilder.setTxPower to send Int8 (1 byte) matching firmware spec
- Fixed doc comment: command code 0x08 → 0x0C, payload 4 bytes → 1 byte
- Updated MeshCoreSession.setTxPower parameter type Int → Int8
- Removed unnecessary Int() cast in SettingsService
- Added tests for positive and negative TX power packet format

+++ MeshCore/Sources/MeshCore/Session/MeshCoreSession.swift
@@ -998,9 +998,9 @@ public actor MeshCoreSession: MeshCoreSessionProtocol {
 
     /// Sets the radio transmission power level.
     ///
-    /// - Parameter power: Power level in dBm (device-specific range, typically 1-20).
+    /// - Parameter power: Power level in dBm (range: -9 to 30).
     /// - Throws: ``MeshCoreError/timeout`` or ``MeshCoreError/deviceError(code:)`` on failure.
-    public func setTxPower(_ power: Int) async throws {
+    public func setTxPower(_ power: Int8) async throws {
         try await sendSimpleCommand(PacketBuilder.setTxPower(power))
     }
